Settings - refactor
