+++ src/page/settings/section/control.rs
@@ -1,0 +1,32 @@
+mod checkbox;
+pub use checkbox::checkbox;
+
+mod label;
+pub use label::label;
+
+mod link_label;
+pub use link_label::link_label;
+
+mod button_label;
+pub use button_label::button_label;
+
+mod connect_button;
+pub use connect_button::connect_button;
+
+mod dropdown;
+pub use dropdown::dropdown;
+
+mod color_picker;
+pub use color_picker::color_picker;
+
+mod user_info;
+pub use user_info::user_info;
+
+mod large_button;
+pub use large_button::large_button;
+
+mod url;
+pub use url::url;
+
+mod status;
+pub use status::status;
