+++ src/page/settings/section/control/button_label.rs
