+++ src/page/settings/section/control/checkbox.rs
