+++ src/page/settings/section/control/label.rs
