+++ src/page/settings/section/control/link_label.rs
