+++ src/page/settings/section/control/url.rs
@@ -1,0 +1,79 @@
+use seed::{prelude::*, *};
+use seed_hooks::{*, topo::nested as view};
+use seed_styles::{em, pc, rem, Style};
+use seed_styles::*;
+use crate::styles::{self, themes::{Color, Breakpoint}, global};
+use crate::page::settings::Msg;
+
+fn on_click_not_implemented() -> EventHandler<Msg> {
+    ev(Ev::Click, |_| { window().alert_with_message("Not implemented!"); })
+}
+
+#[view]
+pub fn url(url: &str, title: &str) -> Node<Msg> {
+    div![
+        C!["option-input-container", "configure-input-container"],
+        s()
+            .padding("0")
+            .align_items(CssAlignItems::Center)
+            .display(CssDisplay::Flex)
+            .flex("1 1 50%")
+            .flex_direction(CssFlexDirection::Row),
+        div![
+            C!["label"],
+            s()
+                .flex_grow("1")
+                .padding("0 1rem")
+                .text_overflow("ellipsis")
+                .white_space(CssWhiteSpace::Pre)
+                .color(Color::SurfaceLight5_90)
+                .flex_basis(CssFlexBasis::Auto)
+                .flex_shrink("1")
+                .line_height(rem(1.5)),
+            attrs!{
+                At::Title => url,
+            },
+            url,
+        ],
+        div![
+            C!["configure-button-container", "button-container"],
+            s()
+                .align_items(CssAlignItems::Center)
+                .background_color(Color::Accent3)
+                .display(CssDisplay::Flex)
+                .flex(CssFlex::None)
+                .flex_direction(CssFlexDirection::Row)
+                .height(rem(3))
+                .justify_content(CssJustifyContent::Center)
+                .width(rem(3))
+                .cursor(CssCursor::Pointer),
+            s()
+                .hover()
+                .background_color(Color::Accent3Light1),
+            attrs!{
+                At::TabIndex => 0,
+                At::Title => title,
+            },
+            on_click_not_implemented(),
+            svg![
+                C!["icon"],
+                s()
+                    .fill(Color::SurfaceLight5_90)
+                    .flex(CssFlex::None)
+                    .height(rem(1))
+                    .margin("0")
+                    .width(rem(1))
+                    .overflow(CssOverflow::Visible),
+                attrs!{
+                    At::ViewBox => "0 0 1043 1024",
+                    At::from("icon") => "ic_settings",
+                },
+                path![
+                    attrs!{
+                        At::D => "M791.492 901.421c-0.137 1.886-0.214 4.085-0.214 6.303 0 14.689 3.414 28.58 9.492 40.924l-0.242-0.544c1.442 2.027 2.306 4.553 2.306 7.281 0 5.548-3.572 10.262-8.542 11.967l-0.089 0.027c-37.735 21.585-81.411 40.158-127.33 53.451l-4.284 1.062c-2.114 1.002-4.593 1.587-7.209 1.587-7.903 0-14.559-5.341-16.556-12.61l-0.028-0.12c-20.88-43.535-64.606-73.060-115.229-73.060-26.819 0-51.703 8.287-72.23 22.44l0.428-0.279c-19.628 13.227-34.808 31.704-43.688 53.426l-0.284 0.786c-3.614 8.734-7.529 11.746-17.769 9.035-51.834-13.272-97.233-31.525-139.449-54.835l3.016 1.527c-14.758-7.831-8.734-16.866-5.12-26.805 4.846-12.398 7.654-26.752 7.654-41.762 0-32.050-12.804-61.11-33.576-82.344l0.021 0.021c-22.874-25.484-55.92-41.441-92.693-41.441-10.83 0-21.336 1.384-31.352 3.985l0.864-0.191h-5.722c-30.118 9.336-30.118 9.035-44.273-18.372-17.236-31.193-32.683-67.512-44.377-105.477l-1.101-4.152c-3.915-12.348-1.807-18.673 11.445-24.094 45.171-18.059 76.501-61.451 76.501-112.16 0-0.275-0.001-0.549-0.003-0.823l0 0.042c-0.157-51.84-32.003-96.203-77.176-114.748l-0.829-0.301c-13.553-4.819-15.962-10.842-12.047-23.793 13.962-48.504 31.914-90.674 54.24-130.036l-1.534 2.94c6.024-10.541 11.746-12.649 23.793-7.831 14.648 6.459 31.727 10.219 49.685 10.219 35.285 0 67.18-14.517 90.038-37.904l0.023-0.024c21.532-21.755 34.835-51.691 34.835-84.733 0-19.022-4.409-37.015-12.26-53.011l0.314 0.709c-4.216-9.638-3.012-15.059 6.024-20.48 39.702-23.013 85.609-42.536 133.977-56.195l4.263-1.029c13.252-3.614 14.758 5.12 18.372 13.252 16.261 41.325 53.282 71.221 97.87 77.036l0.614 0.065c6.241 1.121 13.425 1.762 20.759 1.762 40.852 0 77.059-19.886 99.469-50.507l0.242-0.347c7.452-9.232 13.404-20.047 17.264-31.809l0.204-0.718c3.012-8.433 8.132-9.939 16.264-8.132 52.584 13.65 98.681 32.83 141.232 57.456l-2.691-1.437c9.336 5.12 8.433 11.144 4.819 19.576-6.604 14.774-10.451 32.016-10.451 50.158 0 69.362 56.229 125.591 125.591 125.591 18.623 0 36.299-4.053 52.195-11.326l-0.784 0.321c10.24-4.518 15.962-3.012 21.384 6.927 22.212 37.657 40.917 81.17 53.87 127.095l0.944 3.916c2.711 10.24 0 15.36-10.24 19.878-46.208 16.823-78.61 60.371-78.61 111.487 0 0.299 0.001 0.599 0.003 0.898l-0-0.046c-0.106 1.871-0.166 4.060-0.166 6.264 0 49.766 30.792 92.34 74.362 109.71l0.797 0.28c12.951 6.024 16.264 11.746 12.047 25.6-14.446 47.781-32.562 89.199-54.858 127.907l1.55-2.918c-5.421 10.24-10.842 12.348-22.287 8.132-14.209-5.966-30.724-9.432-48.048-9.432-45.354 0-85.159 23.756-107.651 59.503l-0.31 0.527c-11.029 16.816-17.591 37.422-17.591 59.561 0 1.826 0.045 3.642 0.133 5.446l-0.010-0.254zM520.433 711.68c109.44-1.529 197.571-90.604 197.571-200.264 0-110.613-89.669-200.282-200.282-200.282s-200.282 89.669-200.282 200.282c0 0.205 0 0.411 0.001 0.616l-0-0.032c0.498 110.402 90.11 199.707 200.582 199.707 1.166 0 2.329-0.010 3.49-0.030l-0.175 0.002z",
+                    }
+                ]
+            ]
+        ]
+    ]
+}
